(unsigned long)(1 << 31) = bad news on x86_64. (cherry picked from 410e5b1d738ba47b36778e6cbed44023a27ce259 commit)

+++ fb/fbimage.c
@@ -68,7 +68,7 @@ fbPutImage (DrawablePtr	pDrawable,
 	break;
     case XYPixmap:
 	srcStride = BitmapBytePad(w + leftPad) / sizeof (FbStip);
-	for (i = 1 << (pDrawable->depth - 1); i; i >>= 1)
+	for (i = (unsigned long)1 << (pDrawable->depth - 1); i; i >>= 1)
 	{
 	    if (i & pGC->planemask)
 	    {
